+++ durabletask-azuremanaged/durabletask/azuremanaged/preview/sandboxes/__init__.py
@@ -16,6 +16,7 @@ Usage::
 """
 
 from durabletask.azuremanaged.preview.sandboxes.client import SandboxActivitiesClient
+from durabletask.azuremanaged.preview.sandboxes.helpers import SandboxActivity
 from durabletask.azuremanaged.preview.sandboxes.worker_profiles import SandboxWorkerProfile
 from durabletask.azuremanaged.preview.sandboxes.worker_profiles import SandboxWorkerProfileOptions
 from durabletask.azuremanaged.preview.sandboxes.worker_profiles import sandbox_worker_profile
@@ -23,6 +24,7 @@ from durabletask.azuremanaged.preview.sandboxes.worker import SandboxWorker
 
 __all__ = [
     "SandboxWorker",
+    "SandboxActivity",
     "SandboxWorkerProfile",
     "SandboxWorkerProfileOptions",
     "SandboxActivitiesClient",