+++ examples/sandboxes/activity_names.py
@@ -1,4 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT License.
-
-REMOTE_HELLO = "remote_hello"